refactor: rename AppAppBar to AppNavBar
